[OGWA-21]
fixing bug
