+++ projects/server/src/index.js
@@ -77,7 +77,7 @@ server.listen(PORT, (err) => {
 	if (err) {
 		console.log(`ERROR: ${err}`);
 	} else {
-		db.sequelize.sync({ alter: true });
+		//db.sequelize.sync({ alter: true });
 		console.log(`SERVER IS RUNNING AT PORT:${PORT} ✅`);
 	}
 });
